add: helper and refactoring
