+++ lib/helper/words/retrieve.dart
@@ -1,11 +1,9 @@
 import 'package:notion_wordbook/models/enums/word_tag.dart';
-import 'package:notion_wordbook/models/word.dart';
-import 'package:notion_wordbook/helper/exists.dart';
+import 'package:notion_wordbook/helper/words/exists.dart';
 
 String getSpelling(wordData) {
   return wordData['Spelling']['title'][0]['text']['content'];
 }
-
 
 String getMeaning(wordData) {
   return existsMeaning(wordData)
@@ -36,15 +34,4 @@ List<WordTag> getTags(wordData) {
               (tag) => WordTagHelper().valueOf(tag['name'].toString()))
           .toList()
       : <WordTag>[];
-}
-
-List extractWordsDataFromResponse(Map response) {
-  return response["results"].map((page) => page['properties']).toList();
-}
-
-List<WordModel> makeArray(List wordsData) {
-  return wordsData
-      .map((w) => WordModel(getSpelling(w), getCorrect(w), getMeaning(w),
-          getTags(w), getExampleSentence(w), getLink(w)))
-      .toList();
 }
