--- lib/hooks/hook_test.dart
@@ -1,17 +1,0 @@
-// import 'package:flutter/material.dart'; //google提供のUIデザイン
-// import 'package:http/http.dart' as http; //httpリクエスト用
-// import 'dart:async'; //非同期処理用
-// import 'dart:convert'; //httpレスポンスをJSON形式に変換用
-
-// Future getData() async {
-//   //Future xxx async{} という記法
-//   http.Response response = await http.get("https://reqres.in/api/users?page=2");
-// }
-
-// // 非同期処理は、デフォルトでは呼び出し元は処理の完了を待ちませんが、
-// // await キーワードをつけると完了を待つことができる。
-// @override
-// void initState() {
-//   super.initState();
-//   getData();
-// }
